requirements.txt
gradio
PyPDF2
llama-cpp-python
langchain
sentence-transformers
faiss-cpu
